src/sage/groups/galois_group_perm.py: Split out from src/sage/groups/galois_group.py
